requirements.txt
# Core ML & Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Frameworks
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
streamlit>=1.28.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# ML Model Persistence
joblib>=1.3.0

# Image Processing
Pillow>=10.0.0

# Data Validation
pydantic>=2.0.0

# HTTP Requests
requests>=2.31.0
httpx>=0.24.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Development Tools
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
